Move mempool trait to consensus crate
