+++ consensus/src/consensus.rs
@@ -2,11 +2,10 @@ use crate::config::{Committee, Parameters};
 use crate::core::Core;
 use crate::error::ConsensusResult;
 use crate::leader::LeaderElector;
-use crate::mempool::MempoolDriver;
+use crate::mempool::{MempoolDriver, NodeMempool};
 use crate::messages::Block;
 use crate::synchronizer::Synchronizer;
 use crypto::{PublicKey, SignatureService};
-use mempool::NodeMempool;
 use network::{NetReceiver, NetSender};
 use store::Store;
 use tokio::sync::mpsc::{channel, Sender};
