+++ consensus/src/lib.rs
@@ -17,4 +17,5 @@ mod common;
 pub use crate::config::{Committee, Parameters};
 pub use crate::consensus::Consensus;
 pub use crate::error::ConsensusError;
+pub use crate::mempool::{NodeMempool, PayloadStatus};
 pub use crate::messages::Block;
